requirements.txt
tinygrad
einops
transformers
pillow